requirements.txt
fastapi
uvicorn
openai
fastapi-clerk-auth
PyJWT
cryptography